requirements.txt
# Core AI (Provided by Ultralytics image)
# torch
# torchvision
transformers>=4.35.0
open_clip_torch>=2.24.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
pillow>=10.0.0
numpy>=1.24.0

# Computer Vision
opencv-python-headless>=4.8.0
# yolov5 (Installed manually in Dockerfile to fix setuptools issue)
# git+https://github.com/ultralytics/yolov5.git@master

# Vector Database
# faiss-gpu (Causes build error on Py3.11, using CPU version for now)
faiss-cpu

# Metadata & Files
PyExifTool>=0.5.0
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
pandas>=2.1.0
streamlit>=1.30.0

# YOLOv5 dependencies
huggingface_hub>=0.23.0
seaborn>=0.11.0
thop>=0.1.1
